Fix ISUB overflow tests: move to examples/overflow/ with boundary-forcing guards
Old tests in rangerregression/ had flawed assertions that didn't correctly
validate overflow wrapping. New tests use Verifier.nondetInt() + guard to pin
to boundary values (MAX_VALUE/MIN_VALUE), then simple equality checks.

TRUE: MAX_VALUE - (-1) == MIN_VALUE (holds)
FALSE: MIN_VALUE - 1 != MAX_VALUE (violated)
