+++ jpf-symbc/src/examples/overflow/config-isub.jpf
@@ -1,0 +1,16 @@
+target=overflow.ISUBOverflow_true.MainProp1
+
+classpath=${jpf-symbc}/build/examples
+
+symbolic.dp=z3bitvector 
+symbolic.min_int=-2147483648 
+symbolic.max_int=2147483647 
+symbolic.min_double=-10000.0 
+symbolic.max_double=10000.0 
+symbolic.bvlength=64 
+search.depth_limit=20
+symbolic.lazy=on 
+symbolic.jrarrays=true 
+listener = .symbc.SymbolicListener
+vm.storage.class=nil
+symbolic.optimizechoices=true
